Finish CRUD for customer address

+++ api/src/Service/CustomerAddressService.php
@@ -37,12 +37,12 @@ class CustomerAddressService
         return $this->createAddressArray($address);
     }
 
-    public function editAddress(int $addressId, Request $request): ?array
+    public function editAddress(int $addressId, Request $request, ?string $userEmail = null): ?array
     {
         $address = $this->addressRepository->findOneBy(['id' => $addressId]);
 
-        if (!$address) {
-            return null;;
+        if (!$address || ($userEmail != null && $userEmail != $address->getCustomer()->getEmail())) {
+            return null;
         }
 
         $content = json_decode($request->getContent(), true);
@@ -63,7 +63,19 @@ class CustomerAddressService
             return false;
         }
 
-        $customer = $address->getCustomer();
+        $this->entityManager->remove($address);
+        $this->entityManager->flush();
+
+        return true;
+    }
+
+    public function deleteAddressByCustomer(int $id, string $email): bool
+    {
+        $address = $this->addressRepository->findOneBy(['id' => $id]);
+
+        if (!$address || $address->getCustomer()->getEmail() != $email) {
+            return false;
+        }
 
         $this->entityManager->remove($address);
         $this->entityManager->flush();
@@ -82,9 +94,32 @@ class CustomerAddressService
         return $this->createAddressArray($address);
     }
 
+    public function getCustomerAddress(int $addressId, string $userEmail): ?array
+    {
+        $address = $this->addressRepository->findOneBy(['id' => $addressId]);
+
+        if (!$address || $address->getCustomer()->getEmail() != $userEmail) {
+            return null;
+        }
+
+        return $this->createAddressArray($address);
+    }
+
     public function getCustomerAddresses(Request $request, ?string $userEmail = null): array
     {
         $customerAddresses = [];
+
+        if (!$userEmail) {
+            $userId = $request->get('customerId', 'all');
+            $addresses = $this->addressRepository->findAddressesByCustomerId($userId);
+        } else {
+            $addresses = $this->addressRepository->findCustomerAddresses($userEmail);
+        }
+
+        /** @var CustomerAddress $address */
+        foreach ($addresses as $address) {
+            $customerAddresses[] = $this->createAddressArray($address);
+        }
 
         return $customerAddresses;
     }
@@ -107,6 +142,24 @@ class CustomerAddressService
 
     private function objectCreator(CustomerAddress $address, array $content): ?CustomerAddress
     {
+        foreach ($content as $fieldName => $fieldValue) {
+            if (property_exists(CustomerAddress::class, $fieldName)) {
+                $setterMethod = 'set' . ucfirst($fieldName);
+
+                if ($setterMethod == 'setCustomer') {
+                    $customer = $this->customerRepository->findOneBy(['id' => $fieldValue]);
+
+                    if (!$customer) {
+                        return null;
+                    }
+
+                    $address->setCustomer($customer);
+                } elseif (method_exists($address, $setterMethod)) {
+                    $address->$setterMethod($fieldValue);
+                }
+            }
+        }
+
         return $address;
     }
 }
